refactor(store): move all of the store back to one `mimeReducer`
