+++ browser/store/reducers/index.js
@@ -1,6 +1,3 @@
-import { combineReducers } from 'redux'
-import shapeReducer from './shapeReducer'
+import mimeReducer from './mimeReducer'
 
-// each reducer will track a single piece of the state, ie:
-// state = { template: [everything managed by templateReducer] }
-export default combineReducers({ shapes: shapeReducer })
+export default mimeReducer
